requirements.txt
streamlit
pymongo[srv]
pandas
plotly
python-dotenv